fix: Crashes with MinGW's STDMUTEX_RECURSION_CHECKS feature
MinGW's standard library comes with a useful feature to debug
bugs with non-recursive std::mutexes, this feature can be
enabled by defining STDMUTEX_RECURSION_CHECKS=1.

Enabling this feature is an ABI breakage, since it adds a new
field to std::mutex. Unfortunately, the feature is enabled by
default when NDEBUG is not defined (debug builds).

As a result if tconcurrent was compiled with a different Debug
or Release setting than its reverse-dependencies, some public
classes (such as shared_base) which contain a std::mutex
end up with an incompatible layout.