Fix crash when inlining a try{} block near a Boost.Context switch
In tconcurrent Boost.Context is used to take care of the low-level
handling of stackful coroutine contexts and switching stacks.

Boost.Context's `jump_fcontext` function is implemented in assembly
for each combination of target platform and architecture.
Its job is to save the current state (so that it may be resumed later)
and to start or resume execution of another task.
On i386 win32 this is done by saving registers on the caller's stack
before switching to the new task's stack and continuing execution.

In tconcurrent the function `run_coroutine` is one of the main places
where Boost.Context is used to switch tasks, this is done by calling
boost's `execution_context::operator()`, which will resume execution
of the "context-function" we constructed it with on the new stack.

For coroutines created with the `async_resumable()` function
the "context-function" of the boost execution context is actually
a lambda calling a std::function callback, which is in fact
a `packaged_task` created by the `package()` function, which contains
a `promise_ptr` whose shared_type is the template struct `shared`,
which finally owns a std::function wrapper around the actual coroutine
user code. This coroutine user code is only called through a `package_caller`
wrapper template, itself wrapped in a `try{} catch(...)` block.

The compiler, however, can inline all these template calls in run_coroutine.
As a result of the way Boost.Context is implemented, the compiler is
never made aware of the stack save-and-switch that occurs before
calling the wrapper templates around the couroutine code, so when it
inlines the try{} block surrounding the call to coroutine code,
the compiler sees the result as a single, normal function that:

1. Calls `jump_fcontext`, inlined from `execution_context::operator()`
2. Declares a try{} block, inlined from `shared::operator()`
3. Calls the coroutine's code inside this try block

As it happens, compilers like MinGW implement exception handling mechanisms
such as setjmp-longjmp by registering a "landing pad" at the *start* of any
function that may want to run cleanup or a catch handler during stack-unwind.
The landing pad checks a "call site" number to know what handling code to run.
As a result, the MinGW compiler can generate code that looks like this:

1. Register (at runtime) this function's landing pad on the current stack
2. Save registers on the current stack, and switch to the coroutine's stack
3. Update the "call site" variable to signal we have entered the try block
4. Call the coroutine

If the coroutine code throws an exception, the stack unwinder will call the
personality routine which instructs it that the `run_coroutine` function
has a `catch (...)` block and wants to handle this exception.
When the i386 win32 SjLj runtime calls our landing pad it will use the
information we registered at the start of the function in order to restore,
among other things, the stack pointer we saved.

Of course at the start of the `run_coroutine` function we were still
running on the old stack, so that's what gets restored for the catch block.
At this point it should be pretty obvious that we have a problem: the catch
block and the runtime will keep using the old stack for their own purpose,
overwritting all our registers that Boost.Context saved before switching!

This was observed as a crash only in release mode on 32bit windows builds,
at adress 0 with no stack and all registers set to 0.
We fix the bug by preventing the try{} block from being inlined.