Revert "Fix crash when inlining a try{} block near a Boost.Context switch"
This reverts commit f73c71b68ac9a3cb1dde843f5c6f13bb61ca9372.

After a long (and old) discussion with Tim, we concluded that while this
might fix the issue in this particular case, it is just a fluke.

The problem occurred only with mingw32 with the SjLj exception
mechanism. However, this model cannot work with coroutines because
the chain of landing pads GCC sets up goes across coroutines boundaries.
I don't remember the particular scenarios, but we can always end-up with
the wrong landing pad set up at the wrong time depending on the mood of
the compiler.

The only real fix here would be that boost context also switched the
landing pads along with the rest of the context, as they do with SEH
exception handling.

In the end, we stopped using mingw32 with SjLj and now use the DWARF
exception mechanism which is similar to how exception works on Linux.