stackless_coroutines: fix handling of shared_futures
futures are move-only, while shared_futures are not. We have
co_await(future&&) to force the move because co_await consumes the
future. However, this is not necessary for shared_futures, so it should
be defined as co_await(shared_future&).

future_awaiter tries to factorize the code by storing a & or a &&, and
moves it only if it can in await_ready.